+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
 
 - 🐛(dimail) improve handling of dimail errors on failed mailbox creation #377
 -  💬(frontend) fix group member removal text #382
+-  💬(frontend) fix add mail domain text #382
 
 ## [1.0.2] - 2024-08-30